+++ src/EventGrid/EventGrid.Test/ScenarioTests/TopicTests.cs
@@ -58,5 +58,12 @@ namespace Microsoft.Azure.Commands.EventGrid.Tests.ScenarioTests
         {
             EventGridController.NewInstance.RunPsTest(_logger, "TopicNewKeyTests");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void EventGrid_TopicsInputMappingCreateGetAndDelete()
+        {
+            EventGridController.NewInstance.RunPsTest(_logger, "TopicInputMappingTests");
+        }
     }
 }