+++ src/EventGrid/EventGrid.Test/ScenarioTests/TopicTests.ps1
@@ -248,3 +248,166 @@ function TopicNewKeyTests {
         Remove-ResourceGroup $resourceGroupName
     }
 }
+
+<#
+.SYNOPSIS
+Tests EventGrid Topic Create, Get and List operations with Input mapping
+#>
+function TopicInputMappingTests {
+    # Setup
+    $location = Get-LocationForEventGrid
+    $topicName = Get-TopicName
+    $topicName2 = Get-TopicName
+    $topicName3 = Get-TopicName
+    $topicName4 = Get-TopicName
+    $topicName5 = Get-TopicName
+    $topicName6 = Get-TopicName
+
+    $resourceGroupName = Get-ResourceGroupName
+    $subscriptionId = Get-SubscriptionId
+
+    $expectedInputMappingEventGridSchema = "EventGridSchema"
+    $expectedInputMappingCloudEventSchema = "CloudEventSchemaV1_0"
+    $expectedInputMappingCustomEventSchema = "CustomEventSchema"
+
+    $inputSchemaInvalid = "InvalidSchema"
+    $inputSchemaEventGridSchema1 = "eventgriDSChemA"
+    $inputSchemaEventGridSchema2 = "eventgridschema"
+    $inputSchemaCloudEventSchema1 = "CloUdEvENtSchemaV1_0"
+    $inputSchemaCloudEventSchema2 = "ClOUdEVentSchEMAv1_0"
+    $inputSchemaCustomEventSchema1 = "cUsTomEVeNTSchEma"
+    $inputSchemaCustomEventSchema2 = "customeventschema"
+
+    New-ResourceGroup $resourceGroupName $location
+
+    try
+    {
+        try
+        {
+            Write-Debug "Creating a new EventGrid Topic: $topicName in resource group $resourceGroupName with InputSchema $inputSchemaInvalid"
+            $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName -Location $location -InputSchema $inputSchemaInvalid
+            Assert-True {$false} "New-AzEventGridTopic succeeded while it is expected to fail"
+        }
+        catch
+        {
+            Assert-True {$true}
+        }
+
+        Write-Debug "Creating a new EventGrid Topic: $topicName in resource group $resourceGroupName with InputSchema $inputSchemaEventGridSchema1"
+        $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName -Location $location -InputSchema $inputSchemaEventGridSchema1
+        Assert-True {$result.ProvisioningState -eq "Succeeded"}
+        Assert-True {$result.InputSchema -eq $expectedInputMappingEventGridSchema} "EventGridSchema is expected."
+
+        Write-Debug "Getting the created topic within the resource group"
+        $createdTopic = Get-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName
+        Assert-True {$createdTopic.Count -eq 1}
+        Assert-True {$createdTopic.TopicName -eq $topicName} "Topic created earlier is not found."
+        Assert-True {$createdTopic.InputSchema -eq $expectedInputMappingEventGridSchema} "InputSchema is not correct. EventGridSchema is expected."
+
+        Write-Debug "Creating a second EventGrid topic: $topicName2 in resource group $resourceGroupName with InputSchema $inputSchemaEventGridSchema2"
+        $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName2 -Location $location -Tag @{ Dept = "IT"; Environment = "Test" } -InputSchema $inputSchemaEventGridSchema2
+        Assert-True {$result.ProvisioningState -eq "Succeeded"}
+        Assert-True {$result.InputSchema -eq $expectedInputMappingEventGridSchema} "InputSchema is not correct. EventGridSchema is expected"
+
+        Write-Debug "Creating a third EventGrid topic: $topicName3 in resource group $resourceGroupName with InputSchema $inputSchemaCloudEventSchema1"
+        $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName3 -Location $location -InputSchema $inputSchemaCloudEventSchema1
+        Assert-True {$result.ProvisioningState -eq "Succeeded"}
+        Assert-True {$result.InputSchema -eq $expectedInputMappingCloudEventSchema} "InputSchema is not correct. CloudEventSchema is expected."
+
+        Write-Debug "Creating a fourth EventGrid topic: $topicName4 in resource group $resourceGroupName with InputSchema $inputSchemaCloudEventSchema2"
+        $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName4 -Location $location -InputSchema $inputSchemaCloudEventSchema2
+        Assert-True {$result.ProvisioningState -eq "Succeeded"}
+        Assert-True {$result.InputSchema -eq $expectedInputMappingCloudEventSchema} "InputSchema is not correct. CloudEventSchema is expected."
+
+        Write-Debug "Getting the created topic within the resource group"
+        $createdTopic = Get-AzEventGridTopic -ResourceId "/subscriptions/$subscriptionId/resourceGroups/$resourceGroupName/providers/Microsoft.EventGrid/topics/$topicName3"
+        Assert-True {$createdTopic.Count -eq 1}
+        Assert-True {$createdTopic.TopicName -eq $topicName3} "$topicName3 created earlier is not found."
+
+        Write-Debug "Listing all the topics created in the resourceGroup $resourceGroupName"
+        $allCreatedTopics = Get-AzEventGridTopic -ResourceGroup $resourceGroupName
+        Assert-True {$allCreatedTopics.PsTopicsList.Count -eq 4 } "Topics created earlier is not found in the list."
+
+        Write-Debug "Getting all the topics created in the subscription"
+        $allCreatedTopics = Get-AzEventGridTopic
+        Assert-True {$allCreatedTopics.PsTopicsList.Count -ge 0} "Topics created earlier are not found."
+
+        try
+        {
+            Write-Debug "Creating a fifth EventGrid topic: $topicName5 in resource group $resourceGroupName with InputSchema $inputSchemaEventGridSchema1"
+            $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName5 -Location $location -InputSchema $inputSchemaEventGridSchema1 -InputMappingField @{ id = "MyIdField"; topic = "MyTopicField"; eventtime = "MyEventTimeField"; subject = "MySubjectField"; eventtype = "MyEventTypeField" }
+            Assert-True {$false} "New-AzEventGridTopic succeeded while it is expected to fail as inputmapping parameters is not null for input mapping schema EventGridSchema"
+        }
+        catch
+        {
+            Assert-True {$true}
+        }
+
+        try
+        {
+            Write-Debug "Creating a fifth EventGrid topic: $topicName5 in resource group $resourceGroupName with InputSchema $inputSchemaCloudEventSchema1"
+            $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName5 -Location $location -InputSchema $inputSchemaCloudEventSchema1 -InputMappingField @{ id = "MyIdField"; topic = "MyTopicField"; eventtime = "MyEventTimeField"; subject = "MySubjectField"; eventtype = "MyEventTypeField" }
+            Assert-True {$false} "New-AzEventGridTopic succeeded while it is expected to fail as inputmapping parameters is not null for input mapping schema CloudEventSchema"
+        }
+        catch
+        {
+            Assert-True {$true}
+        }
+
+        try
+        {
+            Write-Debug "Creating a fifth EventGrid topic: $topicName5 in resource group $resourceGroupName with InputSchema $inputSchemaCustomEventSchema1"
+            $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName5 -Location $location -InputSchema $inputSchemaCustomEventSchema1
+            Assert-True {$false} "New-AzEventGridTopic succeeded while it is expected to fail as inputmapping parameters are null"
+        }
+        catch
+        {
+            Assert-True {$true}
+        }
+
+        Write-Debug "Creating a fifth EventGrid topic: $topicName5 in resource group $resourceGroupName with InputSchema $inputSchemaCustomEventSchema1"
+        $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName5 -Location $location -InputSchema $inputSchemaCustomEventSchema1 -InputMappingField @{ id = "MyIdField"; topic = "MyTopicField"; eventtime = "MyEventTimeField"; subject = "MySubjectField"; eventtype = "MyEventTypeField"; dataversion = "MyDataVersionField" }
+        Assert-True {$result.ProvisioningState -eq "Succeeded"}
+        Assert-True {$result.InputSchema -eq $expectedInputMappingCustomEventSchema} "InputSchema is not correct. CustomEventSchema is expected"
+
+        Write-Debug "Creating a sixth EventGrid topic: $topicName6 in resource group $resourceGroupName with InputSchema $inputSchemaCustomEventSchema2"
+        $result = New-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName6 -Location $location -InputSchema $inputSchemaCustomEventSchema2 -InputMappingField @{ id = "MyIdField"; topic = "MyTopicField"; eventtime = "MyEventTimeField"; subject = "MySubjectField"; eventtype = "MyEventTypeField"; dataversion = "MyDataVersionField" } -InputMappingDefaultValue @{ subject = "MySubjectDefaultValue"; eventtype = "MyEventTypeDefaultValue"; dataversion = "MyDataVersionDefaultValue" }
+        Assert-True {$result.ProvisioningState -eq "Succeeded"}
+        Assert-True {$result.InputSchema -eq $expectedInputMappingCustomEventSchema} "InputSchema is not correct. CustomEventSchema is expected"
+
+        Write-Debug "Getting the created topic within the resource group"
+        $createdTopic = Get-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName5
+        Assert-True {$createdTopic.Count -eq 1}
+        Assert-True {$createdTopic.TopicName -eq $topicName5} "Topic created earlier is not found."
+        Assert-True {$createdTopic.InputSchema -eq $expectedInputMappingCustomEventSchema} "InputSchema is not correct. CustomEventSchema is expected"
+
+        Write-Debug "Deleting topic: $topicName"
+        Remove-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName
+
+        Write-Debug "Deleting topic: $topicName2 using the ResourceID parameter set"
+        Remove-AzEventGridTopic -ResourceId "/subscriptions/$subscriptionId/resourceGroups/$resourceGroupName/providers/Microsoft.EventGrid/topics/$topicName2"
+
+        Write-Debug "Deleting topic: $topicName3"
+        Remove-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName3
+
+        Write-Debug "Deleting topic: $topicName4"
+        Remove-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName4
+
+        Write-Debug "Deleting topic: $topicName5"
+        Remove-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName5
+
+        Write-Debug "Deleting topic: $topicName6"
+        Remove-AzEventGridTopic -ResourceGroup $resourceGroupName -Name $topicName6
+
+        # Verify that all topics have been deleted correctly
+        $returnedTopics1 = Get-AzEventGridTopic -ResourceGroup $resourceGroupName
+        Assert-True {$returnedTopics1.PsTopicsList.Count -eq 0}
+
+        $returnedTopics2 = Get-AzEventGridTopic -ResourceGroup $resourceGroupName
+        Assert-True {$returnedTopics2.PsTopicsList.Count -eq 0}
+    }
+    finally
+    {
+        Remove-ResourceGroup $resourceGroupName
+    }
+}