+++ src/EventGrid/EventGrid/ChangeLog.md
@@ -18,6 +18,18 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated to use the 2020-06-01 API version.
+* New features:
+	- Input mapping
+	- Event Delivery Schema
+	- Private Link
+	- Cloud Event V10 Schema
+	- Service Bus Topic As Destination
+	- Azure Function As Destination
+	- WebHook Batching
+	- Secure webhook (AAD support)
+	- IpFiltering
+
 
 ## Version 1.2.3
 * Update references in .psd1 to use relative path